+++ src/platform.ts
@@ -171,6 +171,9 @@ export class Hoffmation implements DynamicPlatformPlugin {
     if (device.deviceCapabilities.includes(DeviceCapability.temperatureSensor) && config.useTemperatureDevices) {
       return true;
     }
+    if (device.deviceCapabilities.includes(DeviceCapability.garageDoorOpener) && config.useGarageDoorDevices) {
+      return true;
+    }
     return false;
   }
 }