+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+## **WORK IN PROGRESS**
+* (theimo1221) Add GarageDoor Opener
 ## 0.3.1 (2023-10-04)
 * (theimo1221) Don't treat cameras as motion sensors